# docs/swagger.yaml
definitions:
  api.ErrorResponse:
    properties:
      error:
        example: An error occurred
        type: string
    type: object
  api.SuccessResponse:
    properties:
      status:
        example: success
        type: string
    type: object
  consts.AgentCampaignStatus:
    enum:
    - draft
    - scheduled
    - sending
    - sent
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - AgentCampaignStatusDraft
    - AgentCampaignStatusScheduled
    - AgentCampaignStatusSending
    - AgentCampaignStatusSent
    - AgentCampaignStatusFailed
    - AgentCampaignStatusCancelled
  dto.AgentCampaignListResponse:
    properties:
      campaigns:
        items:
          $ref: '#/definitions/dto.AgentCampaignResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.AgentCampaignResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      real_sent_count:
        type: integer
      scheduled_at:
        type: string
      status:
        type: string
      target_count:
        type: integer
      target_details:
        items:
          type: string
        type: array
      target_type:
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  dto.AgentMessageListResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/dto.AgentMessageResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      stats:
        $ref: '#/definitions/dto.AgentMessageStats'
      total:
        type: integer
    type: object
  dto.AgentMessageResponse:
    properties:
      agent_campaign_id:
        type: integer
      agent_id:
        type: integer
      content:
        description: 通過關聯取得
        type: string
      global_agent_id:
        description: 通過關聯取得
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      sent_at:
        type: string
      title:
        description: 通過關聯取得
        type: string
    type: object
  dto.AgentMessageStats:
    properties:
      read_count:
        type: integer
      total_count:
        type: integer
      unread_count:
        type: integer
    type: object
...
  dto.BatchDeleteAgentCampaignsRequest:
    properties:
      ids:
        description: 要删除的活动ID列表，至少包含一个有效ID
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
  dto.CreateAgentCampaignRequest:
    properties:
      content:
        type: string
      created_by:
        description: 创建者，最大100个字符
        maxLength: 100
        type: string
      scheduled_at:
        type: string
      status:
        enum:
        - draft
        - scheduled
        type: string
      target_details:
        items:
          type: string
        type: array
      target_type:
        enum:
        - all
        - specific
        - line
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - content
    - created_by
    - status
    - target_type
    - title
    type: object
...
  dto.UpdateAgentCampaignRequest:
    properties:
      content:
        type: string
      id:
        type: integer
      scheduled_at:
        type: string
      status:
        enum:
        - draft
        - scheduled
        type: string
      target_details:
        items:
          type: string
        type: array
      target_type:
        enum:
        - all
        - specific
        - line
        type: string
      title:
        maxLength: 255
        type: string
      updated_by:
        maxLength: 100
        type: string
    required:
    - updated_by
    type: object
...
  entity.AgentCampaign:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        description: 建立者
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      real_sent_count:
        description: 實際發送數量
        type: integer
      scheduled_at:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.AgentCampaignStatus'
        description: draft, scheduled, sending, completed, failed, cancelled
      target_count:
        description: 目標代理數量
        type: integer
      target_details:
        description: 目標詳情 (account列表或line路徑)
        items:
          type: string
        type: array
      target_type:
        description: all, specific, line
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        description: 修改者
        type: string
    type: object
...
info:
  contact:
    email: support@jvdiamondtech.com
    name: API Support
    url: http://www.jvdiamondtech.com/support
  description: 用於管理商戶、玩家和管理員的通知服務
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Notification Service API
  version: "1.0"
paths:
  /api/v1/agent-campaigns:
    delete:
      consumes:
      - application/json
      description: |-
        根據ID列表批量刪除代理訊息活動，支援高效能批量操作

        **請求說明:**
        - ids: 要删除的活動ID數組，必須提供至少一個有效ID
        - 無效ID（不存在或不能删除）會被自動跳過
        - 只有狀態允許删除的活動才會被處理

        **響應說明:**
        - 200: 批量删除成功（部分成功也返回200）
        - 400: 請求參數錯誤
        - 500: 服務器內部錯誤
      parameters:
      - description: 批量刪除請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchDeleteAgentCampaignsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 批量刪除代理訊息活動
      tags:
      - 代理訊息活動
    get:
      description: |-
        分頁列出代理訊息活動，支持多條件篩選
        查詢參數說明：
        - page: 頁碼，從1開始，預設為1
        - page_size: 每頁數量，範圍1-100，預設為10
        - status: 狀態篩選（多選），可選值：draft(草稿)、scheduled(已排程)、sent(已發送)、cancelled(已取消)
        - created_by: 創建者篩選
        - created_start_at: 創建時間區間開始（ISO 8601格式）
        - created_end_at: 創建時間區間結束（ISO 8601格式）
        - scheduled_start_at: 發送時間區間開始（ISO 8601格式）
        - scheduled_end_at: 發送時間區間結束（ISO 8601格式）
      parameters:
      - default: 1
        description: 頁碼
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 每頁數量
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: 狀態篩選（逗號分隔）
        example: '"draft,scheduled,sent"'
        in: query
        name: status
        type: string
      - description: 創建者
        example: '"admin@example.com"'
        in: query
        name: created_by
        type: string
      - description: 創建時間區間開始
        example: '"2024-01-01T00:00:00Z"'
        format: date-time
        in: query
        name: created_start_at
        type: string
      - description: 創建時間區間結束
        example: '"2024-12-31T23:59:59Z"'
        format: date-time
        in: query
        name: created_end_at
        type: string
      - description: 發送時間區間開始
        example: '"2024-01-01T00:00:00Z"'
        format: date-time
        in: query
        name: scheduled_start_at
        type: string
      - description: 發送時間區間結束
        example: '"2024-12-31T23:59:59Z"'
        format: date-time
        in: query
        name: scheduled_end_at
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AgentCampaignListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 列出代理訊息活動
      tags:
      - 代理訊息活動
    post:
      consumes:
      - application/json
      description: |-
        創建新的代理訊息活動，支持全代理、指定代理、代理線等目標類型
        參數說明：
        - status: 訊息狀態，必填，可選值：draft(草稿)、scheduled(排程)
        - target_type: 目標類型，可選值：all(全部代理)、specific(指定代理)、line(代理線)
        - target_details: 目標詳情，當target_type為specific時填入代理帳號陣列，當target_type為line時填入上級代理帳號
        - scheduled_at: 排程時間，根據status決定是否必填
        * 預約發送：status=draft，scheduled_at可帶可不帶
        * 立即發送：status=scheduled，scheduled_at不得帶入
        * 預約發送：status=scheduled，scheduled_at必須帶入

        範例請求（預約發送）：
        ```json
        {
        "title": "Test title",
        "content": "test45678",
        "status": "scheduled",
        "target_type": "all",
        "target_details": [],
        "scheduled_at": "2025-11-11T18:00:00Z",
        "created_by": "winston888"
        }
        ```

        範例請求（立即發送）：
        ```json
        {
        "title": "Immediate Message",
        "content": "Immediate sending",
        "status": "scheduled",
        "target_type": "all",
        "target_details": [],
        "created_by": "winston888"
        }
        ```
      parameters:
      - description: 創建代理訊息活動請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAgentCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.AgentCampaign'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 創建代理訊息活動
      tags:
      - 代理訊息活動
  /api/v1/agent-campaigns/{id}:
    delete:
      description: 刪除指定ID的代理訊息活動
      parameters:
      - description: 代理訊息活動ID - 系統內部唯一識別碼，必須為正整數
        example: 12345
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 刪除代理訊息活動
      tags:
      - 代理訊息活動
    get:
      description: 根據ID獲取代理訊息活動詳情
      parameters:
      - description: 代理訊息活動ID - 系統內部唯一識別碼，必須為正整數
        example: 12345
        format: int64
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AgentCampaign'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 獲取代理訊息活動
      tags:
      - 代理訊息活動
    put:
      consumes:
      - application/json
      description: |-
        更新指定ID的代理訊息活動
        參數說明：
        - status: 訊息狀態，可選，可選值：draft(草稿)、scheduled(排程)
        - target_type: 目標類型，可選值：all(全部代理)、specific(指定代理)、line(代理線)
        - target_details: 目標詳情，當target_type為specific時填入代理帳號陣列，當target_type為line時填入上級代理帳號
        - scheduled_at: 排程時間，根據status決定是否必填
        * 預約發送：status=draft，scheduled_at可帶可不帶
        * 立即發送：status=scheduled，scheduled_at不得帶入
        * 預約發送：status=scheduled，scheduled_at必須帶入

        範例請求（預約發送）：
        ```json
        {
        "title": "Test update",
        "content": "hello mey friend",
        "status": "scheduled",
        "target_type": "specific",
        "target_details": ["fatcat-john01", "fatcat-yoyo"],
        "scheduled_at": "2025-11-11T18:00:00Z",
        "updated_by": "winston888"
        }
        ```

        範例請求（立即發送）：
        ```json
        {
        "title": "Immediate Update",
        "content": "Immediate update message",
        "status": "scheduled",
        "target_type": "all",
        "target_details": [],
        "updated_by": "winston888"
        }
        ```
      parameters:
      - description: 代理訊息活動ID - 系統內部唯一識別碼，必須為正整數
        example: 12345
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新代理訊息活動請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAgentCampaignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AgentCampaign'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新代理訊息活動
      tags:
      - 代理訊息活動
  /api/v1/agents/{global_agent_id}/messages:
    get:
      consumes:
      - application/json
      description: 分頁獲取指定代理的站內信列表，按創建時間由近到遠排序
      parameters:
      - description: 全域代理ID - 跨系統代理唯一識別符，用於標識特定代理
        example: '"agent-123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: global_agent_id
        required: true
        type: string
      - default: 1
        description: 頁碼 - 分頁查詢的第几頁，必須大於0
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 每頁數量 - 每頁返回的記錄數，範圍1-50
        example: 10
        in: query
        maximum: 50
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AgentMessageListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 獲取代理站內信列表
      tags:
      - 代理站內信
  /api/v1/agents/{global_agent_id}/messages/{message_id}/read:
    put:
      consumes:
      - application/json
      description: 將指定代理的指定訊息標記為已讀
      parameters:
      - description: 全域代理ID - 跨系統代理唯一識別符，用於標識特定代理
        example: '"agent-123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: global_agent_id
        required: true
        type: string
      - description: 訊息ID - 系統內部訊息唯一識別碼，必須為正整數
        example: 98765
        format: int64
        in: path
        name: message_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 標記訊息為已讀
      tags:
      - 代理站內信
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: API-Key
    type: apiKey
swagger: "2.0"